# api/openapi.yaml
openapi: "3.1.0"
info:
  title: EarlyPass API
  version: "0.1.0"
  description: |
    Drop-in viral waitlist widget API. Manage campaigns, signups,
    referrals, and analytics programmatically.
  license:
    name: Business Source License 1.1
    url: https://github.com/earlypass/earlypass/blob/main/LICENSE
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.earlypass.app
    description: Production
security:
  - BearerAuth: []
paths:
...
  /api/v1/auth/magic-link:
    post:
      operationId: requestMagicLink
      summary: Request a sign-in code
      description: |
        Sends a 6-digit OTP code to the provided email address and returns a
        `session_token` that must be presented alongside the code when calling
        `POST /api/v1/auth/verify`. This session binding ensures the code can
        only be redeemed by the client that originally requested it.
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkRequest"
      responses:
        "202":
          description: Sign-in code sent (or would be sent if email exists). The `session_token` must be stored and passed to `POST /api/v1/auth/verify`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MagicLinkResponse"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/verify:
    post:
      operationId: verifyMagicLink
      summary: Verify the OTP sign-in code
      description: |
        Verifies the 6-digit code emailed to the user against the `session_token`
        returned by `POST /api/v1/auth/magic-link`. Both must match the same
        auth attempt. On success returns account info and a long-lived API key.
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkVerifyRequest"
      responses:
        "200":
          description: Code verified — returns account info and API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MagicLinkVerifyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/campaigns:
    post:
      operationId: createCampaign
      summary: Create a campaign
      description: |
        Creates a new waitlist campaign linked to the authenticated account.
        Requires a valid account Bearer token.
      tags:
        - campaigns
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignRequest"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/campaigns/{id}/signups:
    parameters:
      - $ref: "#/components/parameters/CampaignID"
    get:
      operationId: listSignups
      summary: List signups for a campaign
      tags:
        - signups
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of signups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createSignup
      summary: Add a signup to a campaign
      description: |
        Registers a new email address on the waitlist. Validates the email,
        runs fraud checks, and optionally credits a referrer.

        Requires the request to originate from the campaign's configured allowed
        origin (Origin header must match). Also requires a valid X-CSRF-Token
        obtained from GET /api/v1/campaigns/{id}/csrf-token.
      tags:
        - signups
      security: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
          description: |
            Single-use CSRF token obtained from GET /api/v1/campaigns/{id}/csrf-token.
            Prevents cross-site request forgery against the signup endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSignupRequest"
      responses:
        "201":
          description: Signup created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signup"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/v1/campaigns/{id}/leaderboard:
    parameters:
      - $ref: "#/components/parameters/CampaignID"
    get:
      operationId: getLeaderboard
      summary: Get the referral leaderboard
      tags:
        - leaderboard
      security: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/campaigns/{id}/csrf-token:
    parameters:
      - $ref: "#/components/parameters/CampaignID"
    get:
      operationId: getCsrfToken
      summary: Get a CSRF token for a campaign
      description: |
        Returns a single-use CSRF token valid for 5 minutes.
        The widget must include this token in the X-CSRF-Token request header
        when calling POST /api/v1/campaigns/{id}/signups to prevent
        cross-site request forgery.
      tags:
        - widget
      security: []
      responses:
        "200":
          description: CSRF token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CSRFTokenResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/invites/{token}:
    get:
      operationId: getInviteToken
      summary: Inspect an invite token
      description: |
        Returns invite token details. Status is "active" if not yet redeemed,
        "redeemed" if already redeemed. Returns 404 if the token does not exist
        or has expired (>30 days).
      tags:
        - invites
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite token details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteTokenResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/invites/{token}/redeem:
    post:
      operationId: redeemInviteToken
      summary: Redeem an invite token
      description: |
        Marks an invite token as redeemed. Idempotent — if already redeemed,
        returns 200 with the existing redemption data. Returns 404 if the token
        does not exist or has expired (>30 days).
      tags:
        - invites
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token redeemed (or already was redeemed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteTokenResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/verify/{token}:
    get:
      operationId: verifyEmail
      summary: Verify an email address
      description: |
        Verifies the email address associated with the token and redirects
        to the campaign's success page (or a default page). No authentication required.
      tags:
        - signups
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to success page
          headers:
            Location:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/w/{slug}:
    get:
      operationId: getWidgetData
      summary: Get widget data for a campaign
      description: |
        Public endpoint used by the embedded widget. Returns campaign info,
        total signup count, and optionally a specific signup's position/referral data.
      tags:
        - widget
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: email
          in: query
          schema:
            type: string
            format: email
        - name: referral_code
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Widget data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetData"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: EarlyPass API Key
      description: |
        Account API key or OAuth access token.
        Pass it as a Bearer token: `Authorization: Bearer ep_acc_...`
        Account API keys are created via magic link auth and scoped to an account.
  parameters:
    CampaignID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Deduplication key for POST/PATCH requests. When provided, the server
        caches the first successful (2xx) response for 24 hours. Subsequent
        requests with the same key, method, and path return the cached response
        with the `Idempotency-Replayed` header set to `true`.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Not authorized to access this resource
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Resource already exists
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnprocessableEntity:
      description: Request is well-formed but semantically invalid (e.g. disposable email)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: "#/components/headers/X-RateLimit-Limit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/X-RateLimit-Remaining"
        Retry-After:
          $ref: "#/components/headers/Retry-After"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    CSRFTokenResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Single-use CSRF token, valid for 5 minutes
...
    Campaign:
      type: object
      required:
        - id
        - name
        - slug
        - status
        - settings
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: My Product Launch
        slug:
          type: string
          example: my-product-launch
        status:
          type: string
          enum: [pending, active, paused, full]
          example: active
        settings:
          $ref: "#/components/schemas/CampaignSettings"
        max_signups:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCampaignRequest:
      type: object
      required:
        - name
        - product_url
      properties:
        name:
          type: string
          maxLength: 200
          example: My Product Launch
        product_url:
          type: string
          format: uri
          example: https://example.com/product
          description: |
            The URL of the product page. Widget API requests are restricted to
            the origin derived from this URL (e.g. https://example.com).
...
    CreateSignupRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        referral_code:
          type: string
          description: Referral code of the person who referred this signup
        fingerprint:
          type: string
          description: Client-supplied fingerprint for fraud analysis
        metadata:
          type: object
          additionalProperties: true
    SignupListResponse:
      type: object
      required:
        - signups
        - total
      properties:
        signups:
          type: array
          items:
            $ref: "#/components/schemas/Signup"
        total:
          type: integer
...
    LeaderboardResponse:
      type: object
      required:
        - entries
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
...
    MagicLinkRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    MagicLinkResponse:
      type: object
      properties:
        message:
          type: string
        session_token:
          type: string
          description: Opaque token that binds this auth attempt to the requesting client. Must be passed to POST /api/v1/auth/verify along with the emailed code.
    MagicLinkVerifyRequest:
      type: object
      required:
        - session_token
        - code
      properties:
        session_token:
          type: string
          description: The session_token returned by POST /api/v1/auth/magic-link.
        code:
          type: string
          description: The 6-digit OTP code received via email.
    MagicLinkVerifyResponse:
      type: object
      properties:
        account:
          $ref: "#/components/schemas/Account"
        api_key:
          type: string
        note:
          type: string
...